Rename trait codegen plugin back to package from module
