+++ examples/trait-package-plugin/create-simple-trait/README.md
