+++ examples/trait-package-plugin/create-simple-trait/build.gradle.kts
@@ -1,7 +1,7 @@
 description = "Custom Smithy structure trait with multiple inputs"
 
 plugins {
-    id("software.amazon.smithy.gradle.smithy-trait-module") version "1.0.0"
+    id("software.amazon.smithy.gradle.smithy-trait-package") version "1.0.0"
 }
 
 group = "software.amazon.smithy"
