+++ examples/trait-package-plugin/create-simple-trait/model/custom-trait.smithy
