+++ examples/trait-package-plugin/create-simple-trait/settings.gradle.kts
