+++ examples/trait-package-plugin/create-simple-trait/smithy-build.json
