+++ examples/trait-package-plugin/use-with-existing-trait/README.md
