+++ examples/trait-package-plugin/use-with-existing-trait/model/custom-trait.smithy
