+++ examples/trait-package-plugin/use-with-existing-trait/settings.gradle.kts
