+++ examples/trait-package-plugin/use-with-existing-trait/smithy-build.json
