+++ examples/trait-package-plugin/use-with-existing-trait/src/main/java/io/smithy/gradle/examples/traits/JsonNameTrait.java
