+++ examples/trait-package-plugin/use-with-existing-trait/src/main/resources/META-INF/services/software.amazon.smithy.model.traits.TraitService
