+++ settings.gradle.kts
@@ -2,5 +2,5 @@ rootProject.name = "smithy-gradle"
 
 include("smithy-base")
 include("smithy-jar")
-include("smithy-trait-module")
+include("smithy-trait-package")
 include("integ-test-utils")
