+++ smithy-trait-package/src/it/java/software/amazon/smithy/gradle/CreatesCustomTraitTest.java
@@ -5,15 +5,13 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.io.File;
 import org.gradle.testkit.runner.BuildResult;
 import org.gradle.testkit.runner.GradleRunner;
-import org.gradle.testkit.runner.TaskOutcome;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-public class RespectsManualTraitTest {
+public class CreatesCustomTraitTest {
 
     @Test
-    public void respectsExistingTraitAndMergesSpiFiles() {
-        Utils.withCopy("trait-module-plugin/use-with-existing-trait", buildDir -> {
+    public void createsTraitsAndAddsToJar() {
+        Utils.withCopy("trait-package-plugin/create-simple-trait", buildDir -> {
             BuildResult result = GradleRunner.create()
                     .forwardOutput()
                     .withProjectDir(buildDir)
@@ -22,11 +20,7 @@ public class RespectsManualTraitTest {
 
             Utils.assertSmithyBuildTaskRan(result);
             Utils.assertValidationRan(result);
-
-            // Check that the merge task was executed successfully
-            Assertions.assertTrue(result.task(":mergeSpiFiles").getOutcome() == TaskOutcome.SUCCESS);
-
-            Utils.assertJarContains(buildDir, "build/libs/use-with-existing-trait-9.9.9.jar",
+            Utils.assertJarContains(buildDir, "build/libs/create-simple-trait-9.9.9.jar",
                     "META-INF/MANIFEST.MF",
                     "META-INF/smithy/manifest",
                     "META-INF/smithy/custom-trait.smithy",
@@ -36,7 +30,7 @@ public class RespectsManualTraitTest {
                     "io/smithy/gradle/examples/traits/JsonNameTrait.class"
             );
 
-            String spiContents = Utils.getJarEntryContents(new File(buildDir, "build/libs/use-with-existing-trait-9.9.9.jar"),
+            String spiContents = Utils.getJarEntryContents(new File(buildDir, "build/libs/create-simple-trait-9.9.9.jar"),
                     "META-INF/services/software.amazon.smithy.model.traits.TraitService");
 
             assertTrue(spiContents.contains("io.smithy.gradle.examples.traits.ResourceMetadataTrait$Provider"));
