+++ smithy-trait-package/src/it/java/software/amazon/smithy/gradle/RespectsManualTraitTest.java
@@ -5,13 +5,15 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.io.File;
 import org.gradle.testkit.runner.BuildResult;
 import org.gradle.testkit.runner.GradleRunner;
+import org.gradle.testkit.runner.TaskOutcome;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-public class CreatesCustomTraitTest {
+public class RespectsManualTraitTest {
 
     @Test
-    public void createsTraitsAndAddsToJar() {
-        Utils.withCopy("trait-module-plugin/create-simple-trait", buildDir -> {
+    public void respectsExistingTraitAndMergesSpiFiles() {
+        Utils.withCopy("trait-package-plugin/use-with-existing-trait", buildDir -> {
             BuildResult result = GradleRunner.create()
                     .forwardOutput()
                     .withProjectDir(buildDir)
@@ -20,7 +22,11 @@ public class CreatesCustomTraitTest {
 
             Utils.assertSmithyBuildTaskRan(result);
             Utils.assertValidationRan(result);
-            Utils.assertJarContains(buildDir, "build/libs/create-simple-trait-9.9.9.jar",
+
+            // Check that the merge task was executed successfully
+            Assertions.assertTrue(result.task(":mergeSpiFiles").getOutcome() == TaskOutcome.SUCCESS);
+
+            Utils.assertJarContains(buildDir, "build/libs/use-with-existing-trait-9.9.9.jar",
                     "META-INF/MANIFEST.MF",
                     "META-INF/smithy/manifest",
                     "META-INF/smithy/custom-trait.smithy",
@@ -30,7 +36,7 @@ public class CreatesCustomTraitTest {
                     "io/smithy/gradle/examples/traits/JsonNameTrait.class"
             );
 
-            String spiContents = Utils.getJarEntryContents(new File(buildDir, "build/libs/create-simple-trait-9.9.9.jar"),
+            String spiContents = Utils.getJarEntryContents(new File(buildDir, "build/libs/use-with-existing-trait-9.9.9.jar"),
                     "META-INF/services/software.amazon.smithy.model.traits.TraitService");
 
             assertTrue(spiContents.contains("io.smithy.gradle.examples.traits.ResourceMetadataTrait$Provider"));
