+++ smithy-trait-package/src/main/java/software/amazon/smithy/gradle/tasks/MergeSpiFilesTask.java
